Исправленное AtHome002

+++ AtHome002/Program.cs
@@ -5,7 +5,7 @@
 // - 3->нет
 // 7->нет
 Console.WriteLine("Введите число: ");
-int num1 = Convert.ToInt32(Console.ReadLine());
+double num1 = Convert.ToInt64(Console.ReadLine());
 if (num1 % 2 == 0)
 {
     Console.WriteLine("Четное");
